requirements.txt
altair==4.0.1
appnope==0.1.0
argh==0.26.2
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
beautifulsoup4==4.8.2
bleach==3.1.0
blinker==1.4
blis==0.4.1
boto3==1.11.7
botocore==1.14.7
Bottleneck==1.3.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
fastai==1.0.60
fastprogress==0.2.2
future==0.18.2
idna==2.8
importlib-metadata==1.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.1.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.3
numexpr==2.7.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
packaging==20.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
pydeck==0.2.0
Pygments==2.5.2
pynvx==1.0.0
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
s3transfer==0.3.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
soupsieve==1.9.5
spacy==2.2.3
srsly==1.0.1
streamlit==0.53.0
terminado==0.8.3
testpath==0.4.4
thinc==7.3.1
toml==0.10.0
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==5.1.1
tqdm==4.41.1
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.8
validators==0.14.1
wasabi==0.6.0
watchdog==0.9.0
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==2.0.0
